Create Oscar_spark_src_2.ligo
This is just the updated Source Code I wrote for Spark Token in Ligo. I want to try translating this back into Michelson, but I'm not too sure of the syntax for creating types in Michelson is. It's very confusing. Check it out and see if it provides any value. I think this looks more friendly than a Michelson script does. I actually commented out the elements of the price of Spark Tokens in Tez and I'm adding a deflationary function to the best of my abilities.